requirements.txt
pycryptodome
SymPy
requests